pyproject.toml
[tool.poetry]
name = "onnx2tensorrt"
version = "0.1.0"
description = "Compile ONNX models to tensorrt engines"
authors = ["Meng Lin <menglinmaker@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
modal = "^0.62.175"
requests = "^2.32.2"
python-dotenv = "^1.0.1"
onnx = "^1.16.1"
boto3 = "^1.34.112"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
